+++ examples/start-end-date.js
@@ -1,7 +1,7 @@
 /* eslint no-console:0 */
 
 import DatePicker from 'antd/lib/date-picker';
-import 'antd/lib/index.css';
+import 'antd/dist/antd.css';
 import createDOMForm from 'rc-form/src/createDOMForm';
 import { regionStyle, errorStyle } from './styles';
 import React, { PropTypes, Component } from 'react';
@@ -57,28 +57,28 @@ class Form extends Component {
       <h2>startTime and endTime validation</h2>
       <form onSubmit={this.onSubmit}>
         <div style={ regionStyle }>
-          <p>start: </p>
-          <p>
+          <div>start: </div>
+          <div>
             <DatePicker {...getFieldProps('start', {
               rules: [this.checkStart],
             })}
             />
-          </p>
+          </div>
         </div>
 
         <div style={ regionStyle }>
-          <p>end: </p>
-          <p>
+          <div>end: </div>
+          <div>
             <DatePicker {...getFieldProps('end', {
               rules: [this.checkEnd],
             })}
             />
-          </p>
+          </div>
         </div>
 
-        <p style={errorStyle}>
+        <div style={errorStyle}>
           {getFieldError('end') ? getFieldError('end').join(',') : ''}
-        </p>
+        </div>
 
         <div style={ regionStyle }>
           <button onClick={this.reset}>reset</button>